refactor: cmd/server から設定パースとFK制約追加処理を別パッケージへ移設
- cmd/server 配下を main.go のみとし、エントリーポイントの責務を起動・DIワイヤリングに限定
- 環境変数パースの純粋関数ヘルパーを internal/app/config パッケージへ移動し公開関数化
- watchlists テーブルのFK制約追加処理を internal/feature/watchlist/adapters/migration.go に移設
- CLAUDE.md のレイアウト記載を新構成に更新
