+++ CLAUDE.md
@@ -116,10 +116,9 @@ feature/<name>/
    - Repositories → Usecases → Handlers のワイヤリングは主に main.go で直接実施
    - `internal/app/di/` には一部のファクトリ関数を配置（例: MarketRepositoryの生成）
 4. **2つのエントリーポイント**:
-   - `cmd/server/`: REST APIサーバー（ポート8080）
-     - `main.go`: 起動・DIワイヤリング
-     - `config.go`: 環境変数パース（`CORS_ALLOWED_ORIGINS` / `COOKIE_SECURE` 等の純粋関数ヘルパー）
-     - `migrations.go`: `AutoMigrate` 後に冪等に実行するFK制約追加処理
+   - `cmd/server/main.go`: REST APIサーバー（ポート8080）の起動・DIワイヤリング
+     - 環境変数パースの純粋関数ヘルパーは `internal/app/config/`（`CORS_ALLOWED_ORIGINS` / `COOKIE_SECURE` 等）
+     - `AutoMigrate` 後のFK制約追加など feature 固有のスキーマ補助は各 feature の `adapters/` に配置（例: `internal/feature/watchlist/adapters/migration.go`）
    - `cmd/ingest/main.go`: TwelveData APIから株価データを取得するバッチジョブ
 
 ### 外部依存
